Move equal trait to matrix functions, introduced function test
