+++ matrix_lib/src/lib.rs
@@ -4,7 +4,6 @@ pub mod errors;
 pub mod dimensions;
 pub mod matrix;
 pub mod matrix_debug;
-pub mod matrix_eq;
 pub mod matrix_convenience;
 pub mod matrix_functions;
 pub mod matrix_modifiers;
